+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* `20/03/27` [opt] publish.gradle
 * `20/01/17` [upd] Leak Canary to v2.1.
 * `19/11/30` [add] Publish bus plugin v2.4. Publish api plugin v1.2.
 * `19/11/28` [add] Publish v1.26.0.
