+++ plugin/api-gradle-plugin/build.gradle
@@ -47,14 +47,14 @@ pluginBundle {
     }
 }
 
-ext {
+apply from: "${rootDir.path}/gradle/publish.gradle"
+publish {
     name = "ApiPlugin"
     groupId = Config.depConfig.plugin_api.groupId
     artifactId = Config.depConfig.plugin_api.artifactId
-    versionName = Config.depConfig.plugin_api.version
-    siteUrl = "https://github.com/Blankj/AndroidUtilCode"
+    version = Config.depConfig.plugin_api.version
+    website = "https://github.com/Blankj/AndroidUtilCode"
 }
-apply from: "${rootDir.path}/gradle/publish.gradle"
 
 //./gradlew plugin:plugin_api-gradle-plugin:mavenLocal     // 上传到本地 mavenLocal
 //./gradlew plugin:plugin_api-gradle-plugin:bintrayUpload  // 上传到 jcenter
